moved parse_region into workflows.basemaps and updated tests
